Move register type enum to register.py
